refactor: reflect response to openapi decleration
