+++ src/polygonParts/models/interfaces.ts
@@ -1,6 +1,7 @@
 import type { PolygonPartsPayload as PolygonPartsPayloadType } from '@map-colonies/mc-model-types';
 import type { PolygonPart, PolygonPartsEntityName, PolygonPartsEntityNameObject, RasterProductTypes } from '@map-colonies/raster-shared';
 import type { MultiPolygon, Polygon } from 'geojson';
+import type { NullableRecordValues } from '../../common/types';
 
 interface CommonPayload extends Omit<PolygonPartsPayload, 'partsData'>, PolygonPart {}
 
@@ -25,12 +26,18 @@ export interface FindPolygonPartsOptions {
 /**
  * Find polygon parts response
  */
-export type FindPolygonPartsResponse = FindPolygonPartsResponseItem[];
+export type FindPolygonPartsResponse = NullableRecordValues<
+  Omit<FindPolygonPartsResponseItem, 'countries' | 'cities' | 'sensors'> & {
+    readonly countries?: string[];
+    readonly cities?: string[];
+    readonly sensors: string[];
+  }
+>[];
 
 /**
  * Find polygon parts response item
  */
-export interface FindPolygonPartsResponseItem extends Omit<PolygonPartRecord, 'partId' | 'insertionOrder'> {}
+export interface FindPolygonPartsResponseItem extends CommonRecord {}
 
 /**
  * Polygon parts ingestion payload
